requirements.txt
PySide6>=6.5.0
pyserial>=3.5
pyautogui>=0.9.54
pygetwindow>=0.0.9
psutil>=5.9.0
